Device ID
On account creation, we are storing the unique device ID of the user. This is for notification purposes. device Id's may also expire, so when a user signs in we check if the current device ID matches what is in firebase, if it does not then we update it there.